Improvements to Style and Structure
In an effort to keep our code style and structure as similar as possible
to the git core project, I've refactored our source to follow some new
guidelines. In essense, I've:
- Migrated core commands to use builtin API
- Made sure line length limits are followed
- Replaced all whitespace with tabs
